[MazaServer] Minor changes
In this change we have now made some minor changes
for this Java 6 application, without a POM file:

1) Package has been corrected and updated for
the project.

2) Gitignore has been updated.

Note: I noticed now that I made some copy and
paste mistakes of the Javadoc descriptions and
forgot to update the text at the same time. This
will not be corrected so excuse the incorrect
and/or invalid descriptions, return values,
parameter values, or even the name of the function
as I feel there is no value in correcting this.